src/backend/GEMINI/requirements.txt
streamlit
sympy
numpy
Pillow
python-dotenv
google-generativeai